pyest_project/requirements.txt
selenium==3.141.0
requests == 2.25.1
PyMySQL==1.0.2
Pytest
pytest-rerunfailures
pytest-xdist
allure-pytest
PyYAML

